Starting new registers' table.
