+++ src/main/java/net/jamsimulator/jams/gui/mips/simulator/registerold/COP0RegisterPropertyWrapper.java
@@ -22,7 +22,7 @@
  *  SOFTWARE.
  */
 
-package net.jamsimulator.jams.gui.mips.simulator.register;
+package net.jamsimulator.jams.gui.mips.simulator.registerold;
 
 import javafx.beans.property.ReadOnlyIntegerProperty;
 import javafx.beans.property.SimpleIntegerProperty;
